app/requirements.txt
streamlit==1.33.0
google-generativeai==0.5.0
python-dotenv==1.0.1
vertexai
geopandas
folium
transformers
langchain==0.3.4
langchain-community==0.3.3
beautifulsoup4==4.12.3
trafilatura==1.12.2
sentence-transformers==3.2.1
faiss-cpu==1.9.0
langchain-openai==0.2.5